Chebyshev-Gauss_Lobatto
Added support for CBL quadrature in the R-Z plane.
Also automated geometry detection for the RBC case,
added option to change start point for Wave number and
energy report output.  Energy report is now integrated
using CGL quadrature. Added a function for global to local
discrete location calculations.  This helps with the for-loops
when data is being translated from local to global.  We don't
need the global to local transform for any of the routines in here.