Condense to one makenek
Removed redudant makenek from repo.  Now we will only use
makenekcyl which is a way to differentiate from standard
makenek scripts.  Also added in copy for PathCreation.py
since it is assumed that when this is run the data will need
to be partitioned.  Actually running PathCreation.py is currently
expected to be taken care of in the submission batch scripts.